Splitting project in several files
